Added media file support
Accepts image, audio and video files and puts them in the proper
locations in the array. If the file type is not supported, an error is
shown and it is forgotten about. Also, when you open the editor, it
takes the data from the arrays and populates the field in a comma list
view.